// ai/image-knowledge-engine/types.ts
export enum ImageType {
  Product = "product",
  Lifestyle = "lifestyle",
  Marketing = "marketing",
  Brand = "brand",
  Packaging = "packaging",
  Banner = "banner",
  Social = "social",
  Catalog = "catalog",
  Other = "other",
}